refactor: promote the generic worker pool out of import_jmap
Pool<J, R> and effective_workers() have no JMAP-specific logic and were
already used by import_dav, import_exchange_graph, import_exchange_ews,
and export, all reaching it through the awkward
crate::sync::import_jmap::pool path. Move it to sync::pool so it reads
as the shared utility it is.
