+++ src/sync/import_dav/items.rs
@@ -17,7 +17,7 @@ use crate::error::Error;
 use crate::jmap::error::JmapError;
 use crate::logging::{LEVEL_PROGRESS, Logger};
 use crate::sync::TypeCounts;
-use crate::sync::import_jmap::pool::Pool;
+use crate::sync::pool::Pool;
 
 use super::calcard;
