+++ src/sync/import_exchange_graph/contacts.rs
@@ -16,7 +16,7 @@ use crate::exchange_graph::contact_map::{ConvertedContact, convert_contact};
 use crate::exchange_graph::error::GraphError;
 use crate::logging::LEVEL_PROGRESS;
 use crate::sync::TypeCounts;
-use crate::sync::import_jmap::pool::Pool;
+use crate::sync::pool::Pool;
 
 use super::coordinator::{CHUNK_SIZE, GraphCoordinator};
 use super::folders::ContactFolder;
